feat(button): added support for disabled theme without disabling button
This is really helpful if you want a button to look disabled without
actually disabling it. A great example for this is creating async
buttons that show a loading progress while some background process or
API call is happening. If the button became `:disabled` during this,
keyboard focus would be lost which isn't ideal.