feat(utils): created a new useResizeObserver implementation
Note: This is fully backwards compatible but the old API will be removed
in the next major version bump.

The new `useResizeObserver` implementation provides a `ref` and a
`refHandler` where the `refHandler` should be passed to the DOM node
that is being watched. The `target` API is kind of deprecated since it
does some weird things and doesn't help with the resize observer loop
behavior.

When using the new API, the event handlers will now use a "subscription"
model where only a single `ResizeObserver` is created instead of multiple
since it helps increase performance.