chore(dev-utils): new release script with pre-compiled theme support
It looks like jsDelivr supports serving files through github which is a
nice workaround for the pre-compiled themes since including the themes
apparently breaks package managers due to the giant size. There will be
documentation in a following commit, but it'll be setup like:

https://cdn.jsdelivr.net/gh/mlaursen/react-md@VERSION/themes/react-md.{PRIMARY}-{SECONDARY}-{SECONDARY_WEIGHT}-{LIGHT|DARK}.min.css

This also fixes an error with the pre-compiled themes for the error
color due to my "amazing" regex replacement to decrease build time.
Since the error color ended up being the same as the first theme color,
the error colors were incorrect.