feat(theme): choose best contrast ratio color by default
This adds a new `function` that is used for default variables within
react-md that tried to guess if the text color should be white or black.
Going forward, this one will use the color that has the biggest contrast
ratio instead of always choosing white if it met the minimum ratio.