fix(table): table border color uses hex values to fix chrome colSpan rendering issue (#982)
The table borders no longer use the `rmd-divider` mixin and colors due
to a weird rendering issue with `colSpan` on tables. The workaround for
this issue is to use a non-rgba value so the table now has new hex
values for this:

- `$rmd-table-light-border-color: #e0e0e0 !default`
- `$rmd-table-dark-border-color: #2f2f2f !default`
- `$rmd-table-border-color: if($rmd-light-theme, $rmd-table-light-border-color, $rmd-table-dark-border-color) !default`

This also added a new `--rmd-table-border-color` CSS variable to
configure this and should automatically update if you use the
`rmd-theme-light/rmd-theme-dark` mixins.