fix(form): Maintain Floating Label for Invalid Numbers
Browsers will try to prevent invalid characters from being entered into
an `<input type="number" />`, but they are unable to prevent numbers
like: `"--00"`, `"0-0"`, or `"123-"`. When these types of numbers are
entered, the input actually reports the value as the empty string
instead of the current text value which caused the floating label to
cover the input when blurred.