chore(sandbox): Removed Sandboxes from Git
The only reason the sandboxes were stored in git was because it took a
long time to compile beforehand. With the new setup, it is fairly quick
and will be part of the normal deploy process instead.

New users will still be able to use the `yarn setup` command to create
an "empty" sandbox lookup, but create the sandboxes on demand with
`yarn sandbox`

These files change too much and are a clutter for reviewing.