fix(list): fixed list icon spacing to work with sass
I'm not sure why, but `sass` renders this as:

```
--rmd-icon-text-spacing: calc(
  var(--rmd-list-text-keyline, 4.5rem)-var(--rmd-list-item-horizontal-padding, 1rem) -
  var(--rmd-icon-size, 1.5rem)
);
```

Since there is no space between the two `var` calls in the calc, it
treats the value as `0`. This update makes it so `sass` renders this
correctly as:

```
--rmd-icon-text-spacing: calc(
  var(--rmd-list-text-keyline, 4.5rem) -
  var(--rmd-list-item-horizontal-padding, 1rem) -
  var(--rmd-icon-size, 1.5rem)
);
```

Note: Added newlines for readability.

I also checked the other usage of `calc` + css variables throughout
react-md, and all of the others seem to work correctly. Super weird.

Closes #1015