refactor(utils): refactored UserInteractionMode hooks and components
This moved all the "logic" into a single hook and added new tests for
this behavior. This also renamed the `InteractionModeListener` to
`UserInteractionModeListener` for more consistency, but still has the
old name available until the next major release. It has been marked as
deprecated which _should_ be caught by editors/IDEs.

> This also reduces the bundle size... by a small amount

```sh
The gizipped UMD bundle sizes are:
 - dist/umd/react-md.production.min.js 97.35 KB (- 1 B)
 - dist/umd/react-md-with-font-icons.production.min.js 120.39 KB (- 4 B)
 - dist/umd/react-md-with-svg-icons.production.min.js 186.74 KB (- 6 B)
```