feat(sass): Updated all packages to simplify sass imports
If you are able to use `sass-loader >= 10`, this change allows you to
import a package's styles with:

```diff
-@import '~@react-md/dist/mixins';
+@import '@react-md/utils';
```

or

```diff
-@import '~@react-md/theme/dist/color-palette';
+@import '@react-md/theme';
```

There are some performance concerns with the second approach if you only
need the color palette and not the full theme package though. This might
be solved in later commits once react-md is updated to use the new
module system.