chore(docs): Fixed weird dev error with spreading props and key
I ended up finding this error since trying to load the TextFields page
crashes in dev mode because an error was thrown about `<textarea>`
should not have multiple `children` and you should use `value` or
`defaultValue` instead. It turns out that some version of `next` makes
it so that if you use the spread operator **before** a `key` on a JSX
element, the element will be forced to have children of:

```ts
[undefined, undefined, false, undefined]
```

As I started writing this commit message, looks like it has been logged
in the nextjs repo and is related to SWC. Should hopefully be fixed
soon.