fix(tooltip): Tooltips stay visible on mobile Firefox
It looks like mobile Firefox triggers a touchmove event around the same
time as a contextmenu event which caused the hide() function to trigger.
Changing it to a scroll event instead of touchmove fixes the behavior so
that a user touching a tooltipped element and starts scrolling the page
does not cause the tooltip to appear.