fix(app-bar): AppBarTitle now applies `flex: 1 1 auto`
This shouldn't really affect most users in any way since this only was
discovered with my documentation site's search behavior on mobile.

If you apply a `max-width` to an element and have `margin-left: auto`,
it won't always be displayed as `flex-end` which was the original
intention.

A side effect of this change is that you won't need to enable the
`first` prop on the first `AppBarAction` if there is an `AppBarTitle` in
your `AppBar`.