chore(transition): fix unmounting transitions
I'm not sure if this is because of server components or something else
but I started noticing that dropdown menus would sometimes flash while
exiting. After debugging, I found out it's because the `className` was
undefined while the `rendered` state was still `true` for a few
milliseconds. So instead of waiting for the useEffect to run one more
time, it's better to immediately unmount instead of going
`exited -> unmount`