chore(snackbar): fix some issues
Render the snackbar only when at least one toast exists. No reason to
have it always mounted since it no longer has a `role="status"`.

Move the `role="status"` to each `Toast` instead of the `Snackbar`
itself. The toast will be the one to render as `status` or `alert`.

The `DefaultToastRenderer` will correctly add close buttons if the toast
itself added `closeButtonProps` instead of globally only.

The `Toast` will correctly override the `closeIcon` if `children` were
passed in the `closeButtonProps`.