feat(core): use :focus-visible instead of keyboard-only :focus
I'm hoping to eventually get rid of all the interaction mode stuff since
CSS is getting more powerful. :focus-visible is _almost_ the same as the
keyboard mode, but not as nice if you switch from mouse to keyboard
without a tab keypress.

Some of the :focus-visible styles require the :has selector which
arrives in Firefox 121. It should be released before I release react-md,
so going to add it in.

The keyboard-only behavior can still be used by setting
`$disable-focus-visible` and the has selectors can be disabled with
`$disable-has-selectors`.