chore(core): remove Typography styling from SrOnly
Since the SrOnly has been updated to also work as a responsive text
container which only appears on tablet and above, the typography should
be removed so the text inherits the parent typography instead. i.e.
buttons.

If typography is desired, just import the Typography component as a
wrapper or use the typography class name utility