chore(form): added tests and fixes for Select
Some additional notes around changes:

I removed the auto-selecting values while closed behavior since it
appears to reduce accessibility with too much noise for screen readers.
The listbox will be shown with ArrowUp/ArrowDown instead of changing the
value and removed the Home/End key behavior while closed.

If I want to mimic the w3.org behavior, search characters should open
the listbox and be selected. I didn't implement that yet and not sure if
I will. I'm working on a `useCombobox` hook to help with auto completes
and debating this part still.