fix(select): allow Select to be autofilled
It seems like browsers won't autofill an input that has
`pointer-events: none`, so removed those styles to allow it to be
autofilled. It doesn't really seem to be an issue since the input will
never really be focused since clicking will open the menu and focus the
combobox once closed instead of the input. It still is not part of the
normal tabindex, so should be good.

This also reused some existing components/styles to style the selected
value/placeholder.