chore(dialog): always use dialog container to prevent unmounting
If you use JS to switch between centered and full page modals, it'll
cause any internal state to be removed which provides bad UX. The dialog
container will be an unstyled div unless the type is set to "centered".

I found out due to the pattern where we use a full page dialog on
smaller viewports and switch to a large centered dialog on bigger
displays. I'll also be adding support for a `responsiveFullPage` with
breakpoints to better support this pattern.